// frontend/src/presentation/components/Button/type.ts
export enum VariantButton {
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
  TERTIARY = 'tertiary',
  QUATERNARY = 'quaternary',
  OUTLINE = 'outline',
  LINK = 'link',
  TEXT = 'text',
  SUCCESS = 'success',
  DANGER = 'danger',
}

export enum TypeButton {
  BUTTON = 'button',
  SUBMIT = 'submit',
}

export enum SizeButton {
  SMALL = 'small',
  MEDIUM = 'medium',
  LARGE = 'large',
}